debian/control: set dependency to python-snimpy
currently the runtime dependency snimpy is set. This is wrong, because
the executable is not used. r4dd uses the snimpy python modules. They
are packaged in 'python-snimpy'.

snimpy depends on python-snimpy on jessie, however in stretch it depends
on python3-snimpy. So r4dd (running as python2 application) is missing
the runtime dependencies on newer systems.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Holger Dengler <dengler@linutronix.de>