r4dd: Add Filehandler creator
The use of streams or filenames in basicConfig() is allowed if only this
one stream is used. If more than one type of output is required,
handlers are necessary.

The logfile option in r4dd describes the creation path to a logfile.
Since the specified path is freely assignable by the user, the
input must be sanitized before the file handler is created.

Signed-off-by: Anton Gres <dalo@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>